requirements.txt
numpy
pytorch>=0.4
torchvision